refactor: change SurveyItemEmpty name to SurveyListEmpty
