+++ src/presentation/pages/SurveyList/components/SurveyListEmpty/SurveyListEmpty.styles.scss
